Eliminate Hamster::Tuple
While I like the concept of "tuples", Hamster::Tuple does not offer
anything over a simple frozen Array. To boot, Array#inspect returns a
String which can be eval'd, whereas Tuple#inspect does not. Using a
frozen Array is idiomatic for Ruby. As much as we would like Ruby to
be more like (insert favorite functional language here), it is what it
is, and Hamster will be more useful if it complies with the
expectations of "normal Ruby programmers".