Set#sort returns a SortedSet
Closes #34.

Note that 'alias' is used rather than 'def_delegator'. This is intentional.
Although internally, the implementation of #sort and #sort_by is the same,
externally, library users should think of them as 2 different API methods.
If one is overridden, there is no reason for the other to change at the
same time (unless the user specifically makes it so).