Remove unneeded indirection from core_ext/enumerable.rb
This file exhibited a common Rails anti-pattern: defining a module for the sole
purpose of including it in a single class, right after the definition. This does
not serve any good purpose. More than that, rather than just defining #to_list
in the module, it used an 'included' hook to define it directly on the
including class.