Fix Hash#hash spec which was based on faulty premise
One of the Hash#hash specs was based on the idea that :issued_at and :expires_at
would hash to the same bucket. That may have been true in an old version of Ruby,
but in up-to-date versions, the hash codes used by Symbols and other core classes
are different on each run.