List#find_indices doesn't blow stack on large lists which don't contain matching items
The optimization applied in 18805a1 means that it is no longer safe for Stream.new
blocks to use recursive calls, *unless* they add at least something to the list
as they do so; in other words, unless the recursive call is done to get the *tail*
of a Sequence.

List#find_indices was safe against stack overflow before, but after applying that
optimization, it was not any more. Now it is safe against stack overflow again.