Approx 50% speed boost for List#union
The implementation of List#union is almost the same as List#uniq now. I don't
really like the extra argument used to pass along the state of the lazy
computation. I would like to come up with a general way of creating lazy lists
where some additional state has to be carried along with the lazy list and
passed down to the next node which must be realized.