List#[] (and #slice) can take a single index, (from,length) args, OR a Range
The return values mimic those of an equivalent Array. This means that if 'from'
is completely out of range, or 'length' is negative, nil is returned. This might
not be very useful in some cases. Then again, it might help to catch bugs:
there is no logical answer with a negative 'length', and returning nil in such
a case makes it likely that a NoMethodError will be raised shortly after.