Use #add instead of redundant #conj/#conjoin
Hamster collection. It was used in #group_by_with, which allows a Vector to be
grouped into multiple Vectors, a Set to be grouped into multiple Sets, and so on.

Part of the idea of #conj was that it would add an item in the most efficient way for
any collection. For Lists, that means cons'ing the new item onto the front.

In the interests of reducing interface bloat, and after discussion on GitHub,
Clojure, #add is a simpler and easier-to-understand name.

The one caveat is that #add was already used in List for adding an item to the
*end* of a List. So this is a backward-compatibility-breaking change.