Replace use of 'def_delgator :self' with ordinary 'alias'
Originally, this library made heavy use of 'def_delegator :self' to create method
aliases. The idea was that if the primary method was redefined, all the aliases would
automatically use the new definition. However, the library documentation never specified
which was the 'primary' method name, which subclasses should redefine. Method calls
which go through an alias are also far slower using 'def_delegator'. Additionally, the
code is not as idiomatic and harder to understand for those unfamiliar with the
'def_delegator :self' pattern.

If subclassers wish to redefine a method along with all its aliases, they can simply
put their own 'alias' lines after the redefinition.