Remove redundant List#find_indices and #elem_indices methods
List#indices already does everything which #find_indices and #elem_indices can do.
This implementation of #indices is a bit clumsy, especially the way it passes
arguments on recursive calls. Maybe it can be improved later.

All the specs for #find_indices and #elem_indices have been combined in a single
spec file for #indices.