Fix race condition in Partitioned#realize
Each partitioned lazy list uses a mutex to prevent more than one thread from trying
to realize the list at the same time. After the list is realized, the mutex isn't
needed any more, so it would be nice to allow it to be GC'd.

The problem is that when @mutex is nulled out, another thread may already have got
into #realize, but not yet read the value of @mutex. Then a NullPointerException
could occur when called @mutex.synchronize.

Therefore, read the value of @mutex into a local variable and check whether it is
nil before using it. If it is nil, that means another thread has already got ahead
and realized the list node, so we can just exit #realize.

Thanks to Ilya Lavrov for finding this problem.