Ensure Set#sort_by calls passed block only once per item
Of course, if items are later added or removed, or inclusion tests are
done in the resulting SortedSet, the block will be called again. So using a
block with side effects with Set#sort_by must be done with care. However, it
is still worthwhile ensuring that the block is called only once per item when
initializing the SortedSet.

Thanks to Rafael Coutinho for reporting this issue.

Fixes GitHub issue #179.