Vector#flatten works on larger, non-nested Vectors initialized with .new
This bug was discovered while investigating GH issue #182, reported by
Xavier Shay. The bug was caused by misunderstanding the behavior of Array#flatten!
It returns nil if no changes were made.

At the same time as this bug is fixed, we can also take advantage of this behavior
to reduce unnecessary object allocations.