Allow Immutable methods to be hidden by subclasses.
My motivation for this is I don't want to leak hamster methods in the
API of one of my libraries. The easiest way to do this is to subclass
Vector and mark as protected any methods that aren't provided by Array.
(I also need to deal with methods that have different return types, but
that's a separate issue.)

Admittedly this use-case is kind of weird - I'm experimenting with a few
different library ideas. This change seemed pretty innocuous though.