Add Rake task to draw graph of load dependencies between files
I've found that sometimes we get errors when only a single file in the Hamster library
is loaded, due to weird issues with the order that files are loaded in.

I'd like to untangle the load order issues so everything works fine no matter what
subset of the library is loaded. These dependency graphs will help with that.