Modernize rspec configuration
Added a modern RSpec configuration to `.rspec` and `spec_helper.rb`,
based on the one generated by `rspec --init`.

This includes quite a number of options that for backwards-compatibility
reasons are not enabled by default, but that add extra checks and that
will be the default on RSpec 4.

See
<https://ivoanjo.me/blog/2016/09/25/why-you-should-regenerate-your-spec-helper/>
for more notes on this.