Use sorted_set gem to implement Hamster.{from,to_ruby} on Ruby 3.0
Ruby's SortedSet used to be included in set.rb (standard library).
Starting from Ruby 3.0, it's not. Rather, it has been broken out into a
gem.

So to keep the one part of Hamster which references SortedSet
working, we need to pull in that gem. I don't like this, but it's
better than potentially breaking user code.